% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zipcarFunctions_function.R
\name{crosstab}
\alias{crosstab}
\title{Crosstab}
\usage{
crosstab(..., dec.places = NULL, type = NULL, style = "wide",
  row.vars = NULL, col.vars = NULL, percentages = TRUE,
  addmargins = TRUE, subtotals = TRUE)
}
\description{
This is a reproduction of the crosstab function created by Dr. Paul Williamson
crosstab()
frequency' for multi-dimenstional tables
}
