% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zipcarFunctions_function.R
\name{detachAllPackages}
\alias{detachAllPackages}
\title{Detach All Packages}
\usage{
detachAllPackages()
}
\description{
This function will unload all previously loaded packages, with the exception(s) of base packages
detachAllPackages()
}
