% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zipcarFunctions_function.R
\name{getMemberClosestPodLocationData}
\alias{getMemberClosestPodLocationData}
\title{Load Member Closest Pod Data files}
\usage{
getMemberClosestPodLocationData(my.file, my.path)
}
\arguments{
\item{my.file}{the name of the file to load.  Required.
getMemberClosestPodLocationData()}

\item{my.path}{the directory path to load the file.  Default path is ../../Data/geo/}
}
\description{
This function leverages the readr package to read and properly format Member Closest Pod Location specific data files
}
