% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zipcarFunctions_function.R
\name{getMemberData}
\alias{getMemberData}
\title{Load Member Data files}
\usage{
getMemberData(my.file, my.path)
}
\arguments{
\item{my.file}{the name of the file to load.  Required.
getMemberData()}

\item{my.path}{the directory path to load the file.  Default path is ../../Data/Members/}
}
\description{
This function leverages the readr package to read and properly format Member specific data files
}
