% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zipcarFunctions_function.R
\name{getMemberSurvivalData}
\alias{getMemberSurvivalData}
\title{Load Member Survival Data files}
\usage{
getMemberSurvivalData(my.file, my.path)
}
\arguments{
\item{my.file}{the name of the file to load.  Required.
getMemberSurvivalData()}

\item{my.path}{the directory path to load the file.  Default path is ../../Data/Survival/}
}
\description{
This function leverages the readr package to read and properly format Member Survival specific data files
}
