% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zipcarFunctions_function.R
\name{getPodLocationData}
\alias{getPodLocationData}
\title{Load Pod Data files}
\usage{
getPodLocationData(my.file, my.path)
}
\arguments{
\item{my.file}{the name of the file to load.  Required.
getPodLocationData()}

\item{my.path}{the directory path to load the file.  Default path is ../../Data/Pod Locations/}
}
\description{
This function leverages the readr package to read and properly format Pod Locations specific data files
}
