% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zipcarFunctions_function.R
\name{getReservationData}
\alias{getReservationData}
\title{Load Reservation Data files}
\usage{
getReservationData(my.file, my.path)
}
\arguments{
\item{my.file}{the name of the file to load.  Required.
getReservationData()}

\item{my.path}{the directory path to load the file.  Default path is ../../Data/Reservations/}
}
\description{
This function leverages the readr package to read and properly format Reservation specific data files
}
