% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zipcarFunctions_function.R
\name{print.crosstab}
\alias{print.crosstab}
\title{Print Crosstab}
\usage{
\method{print}{crosstab}(x, dec.places = x$dec.places,
  subtotals = x$subtotals, ...)
}
\description{
This is a reproduction of the print.crosstab function created by Dr. Paul Williamson
print.crosstab()
}
